Move path mapping setup into pydevd
